+++ ui/Screen/DesignSystemGuide/IconSwatch.svelte
@@ -1,0 +1,12 @@
+<style>
+  .swatch {
+    display: grid;
+    grid-template-columns: repeat(10, 1.5rem);
+    grid-gap: 1.5rem;
+    margin-bottom: 34px;
+  }
+</style>
+
+<div class="swatch">
+  <slot />
+</div>